+++ lib/src/services/client_security.dart
@@ -1,31 +1,34 @@
-// This package provides an abstract class for client authentication and security using certificates.
 import 'dart:io';
 import 'package:flutter/services.dart';
+import 'package:pix_sicoob/src/errors/pix_exception_interface.dart';
 import 'package:result_dart/result_dart.dart';
 
-import '../error/pix_error.dart';
-
+/// An abstract class that provides methods for client security management.
 abstract class ClientSecurity {
-  /// Returns a [SecurityContext] object for secure communication with a server,
-  /// given a base64-encoded certificate string and its password.
+  /// Returns a [Result] containing a [SecurityContext] if successful, or a [PixException] if an error occurred.
   ///
-  /// Throws a [PixError] if the certificate is invalid or the password is incorrect.
-  Result<SecurityContext, PixError> getContext({
+  /// The [certificateBase64String] and [certificatePassword] parameters are used to create a [SecurityContext] object that can be used to communicate securely with the server.
+  ///
+  /// Throws a [PixException] if an error occurred.
+  Result<SecurityContext, PixException> getContext({
     required String certificateBase64String,
     required String certificatePassword,
   });
 
-  /// Reads a PKCS12 certificate file from [pkcs12CertificateFile] and returns its
-  /// base64-encoded string representation, for use in secure communication.
+  /// Returns a [Result] containing a base64 encoded certificate string if successful, or a [PixException] if an error occurred.
   ///
-  /// Throws a [PixError] if the file is not found, is not a valid PKCS12 file,
-  /// or cannot be read for any reason.
-  String certFileToBase64String({required File pkcs12CertificateFile});
+  /// The [pkcs12CertificateFile] parameter is used to convert the PKCS12 certificate file to a base64 encoded string.
+  ///
+  /// Throws a [PixException] if an error occurred.
+  Result<String, PixException> certFileToBase64String(
+      {required File pkcs12CertificateFile});
 
-  /// Converts a base64-encoded certificate string to a byte array, for use in secure communication.
+  /// Returns a [Result] containing a [Uint8List] of bytes representing the certificate string if successful, or a [PixException] if an error occurred.
   ///
-  /// Returns a [Uint8List] object that contains the byte array, or a [PixError] if the string is invalid.
-  Result<Uint8List, PixError> certificateStringToBytes(
+  /// The [certificateBase64String] parameter is used to convert the base64 encoded string to bytes.
+  ///
+  /// Throws a [PixException] if an error occurred.
+  Result<Uint8List, PixException> certificateStringToBytes(
     String certificateBase64String,
   );
 }