+++ lib/src/services/sicoob_client.dart
@@ -1,32 +1,48 @@
-// This class implements a client service using the Sicoob API, with client authentication and security using certificates.
-
 import 'dart:convert';
 import 'dart:io';
 import 'package:http/io_client.dart';
+import 'package:pix_sicoob/src/errors/pix_exception_interface.dart';
+import 'package:pix_sicoob/src/errors/sicoob_http_exception.dart';
 import 'package:pix_sicoob/src/services/client_service.dart';
 import 'package:result_dart/result_dart.dart';
 
+/// A Sicoob implementation of the [ClientService] interface, which provides methods for making HTTP GET and POST requests.
+///
+/// This implementation uses an instance of [IOClient] from the `http` package to make HTTP requests with an [HttpClient]
+/// configured with a [SecurityContext]. Responses are parsed as JSON and returned as a [Map] of key-value pairs.
+///
+/// Example usage:
+/// ```dart
+/// final sicoobClient = SicoobClient(securityContext);
+/// final result = await sicoobClient.get(Uri.parse('https://example.com'));
+/// if (result.isSuccess) {
+///   final responseBody = result.successValue;
+///   // process responseBody
+/// } else {
+///   final exception = result.failureValue;
+///   // handle exception
+/// }
+/// ```
 class SicoobClient implements ClientService {
-  /// Inject a [SecurityContext] with a valid pkcs12 certificate
   final SecurityContext securityContext;
   late HttpClient httpClient;
   late IOClient ioClient;
 
-  /// Constructs a new instance of [SicoobClient] with the given [securityContext].
+  /// Creates an instance of [SicoobClient] with the specified [securityContext] for making secure HTTP requests.
   ///
-  /// Creates an [HttpClient] instance with the given [securityContext], and an [IOClient] instance
-  /// using that [httpClient].
+  /// The [securityContext] must be initialized with the appropriate certificate and key data for the desired server.
   SicoobClient(this.securityContext) {
     httpClient = HttpClient(context: securityContext);
     ioClient = IOClient(httpClient);
   }
 
+  @override
+
   /// Sends an HTTP GET request to the specified [uri], with optional [headers] and [queryParameters].
   ///
-  /// Returns a [Future] that completes with a [Result] object that contains a map of response data,
-  /// or an [Exception] if the request fails.
-  @override
-  Future<Result<Map<String, dynamic>, Exception>> get(
+  /// Returns a [Result] object containing a [Map] of key-value pairs representing the JSON response body if the request is successful.
+  /// Otherwise, returns a [Failure] containing a [PixException] indicating the reason for the failure.
+  Future<Result<Map<String, dynamic>, PixException>> get(
     Uri uri, {
     Map<String, String>? headers,
     Map<String, String>? queryParameters,
@@ -39,17 +55,18 @@ class SicoobClient implements ClientService {
       );
       var jsonBody = jsonDecode(response.body);
       return Success(jsonBody);
-    } on Exception catch (e) {
-      return Failure(e);
+    } catch (e) {
+      return Failure(SicoobHttpException.httpException(e));
     }
   }
 
+  @override
+
   /// Sends an HTTP POST request to the specified [uri], with optional [headers] and [body].
   ///
-  /// Returns a [Future] that completes with a [Result] object that contains a map of response data,
-  /// or an [Exception] if the request fails.
-  @override
-  Future<Result<Map<String, dynamic>, Exception>> post(
+  /// Returns a [Result] object containing a [Map] of key-value pairs representing the JSON response body if the request is successful.
+  /// Otherwise, returns a [Failure] containing a [PixException] indicating the reason for the failure.
+  Future<Result<Map<String, dynamic>, PixException>> post(
     Uri uri, {
     Map<String, String>? headers,
     dynamic body,
@@ -62,8 +79,8 @@ class SicoobClient implements ClientService {
       );
       var jsonBody = jsonDecode(response.body);
       return Success(jsonBody);
-    } on Exception catch (e) {
-      return Failure(e);
+    } catch (e) {
+      return Failure(SicoobHttpException.httpException(e));
     }
   }
 }