+++ lib/src/services/sicoob_security.dart
@@ -1,35 +1,25 @@
 import 'dart:convert';
 import 'dart:io';
 import 'package:flutter/services.dart';
-import 'package:pix_sicoob/src/error/pix_error.dart';
+import 'package:pix_sicoob/src/errors/sicoob_certificate_exception.dart';
+import 'package:pix_sicoob/src/errors/sicoob_unknown_exception.dart';
 import 'package:pix_sicoob/src/services/client_security.dart';
 import 'package:result_dart/result_dart.dart';
 
-///This class provides an implementation of the abstract ClientSecurity class,
-///for client authentication and security using certificates specific to the Sicoob system.
-///
-///The class has three methods:
-///
-/// - `getContext` - Returns a SecurityContext object for secure communication with a server,
-/// given a base64-encoded certificate string and its password.
-/// Throws a PixError if the certificate is invalid or the password is incorrect.
-///
-/// - `certFileToBase64String` - Reads a PKCS12 certificate file and returns
-/// its base64-encoded string representation, for use in secure communication.
-/// Throws a PixError if the file is not found, is not a valid PKCS12 file, or cannot be read for any reason.
-///
-/// - `certificateStringToBytes` - Converts a base64-encoded certificate string to a byte array,
-/// for use in secure communication. Returns a Result object that contains the byte array,
-/// or a PixError if the string is invalid.
+import '../errors/pix_exception_interface.dart';
+
+/// This class implements the [ClientSecurity] interface and provides methods
+/// for managing and retrieving security context and certificates.
 class SicoobSecurity implements ClientSecurity {
-  /// Returns a [SecurityContext] object for secure communication with a server,
+  /// Returns a [Result] object containing a [SecurityContext] object on success
+  /// or a [PixException] on failure. The [SecurityContext] object is created with
+  /// the certificate chain and private key extracted from the given certificate
+  /// base64 string and password.
   ///
-  /// given a base64-encoded certificate string and its password.
-  ///
-  /// Throws a [PixError] if the certificate is invalid or the password is incorrect.
-  ///
+  /// Throws a [SicoobCertificateException] if there is an issue with the certificate,
+  /// or a [SicoobUnknownException] for any other type of error.
   @override
-  Result<SecurityContext, PixError> getContext({
+  Result<SecurityContext, PixException> getContext({
     required String certificateBase64String,
     required String certificatePassword,
   }) {
@@ -48,60 +38,54 @@ class SicoobSecurity implements ClientSecurity {
         );
       return Success(securityContext);
     } on TlsException catch (e) {
-      if (e.osError.toString().contains('INCORRECT_PASSWORD')) {
-        return Failure(PixError('incorrect_certificate_password'));
-      } else if (e.osError.toString().contains('tBAD_PKCS12_DATA'[4])) {
-        return Failure(PixError('invalid_pkcs12_certificate'));
-      } else {
-        return Failure(PixError(e.toString()));
-      }
+      final error = SicoobCertificateException.tlsException(e);
+      return Failure(error);
+    } catch (e) {
+      final unknownError = SicoobUnknownException.unknownException(e);
+      return Failure(unknownError);
     }
   }
 
-  /// Reads a PKCS12 certificate file from [pkcs12CertificateFile] and returns its
-  /// base64-encoded string representation, for use in secure communication.
+  /// Returns a [Result] object containing a base64-encoded string on success or
+  /// a [PixException] on failure. The certificate file is read from the given file
+  /// path, converted to a byte array, and then encoded as a base64 string.
   ///
-  /// Example usage:
-  ///
-  /// ```dart
-  /// final security = SicoobSecurity();
-  /// final certFile = File('path/to/certificate.p12');
-  /// final certString = security.certFileToBase64String(pkcs12CertificateFile: certFile);
-  /// // Use the base64-encoded string for secure communication
-  /// ```
-
+  /// Throws a [SicoobCertificateException] if there is an issue with the certificate,
+  /// or a [SicoobUnknownException] for any other type of error.
   @override
-  String certFileToBase64String({required File pkcs12CertificateFile}) {
-    final bytes = pkcs12CertificateFile.readAsBytesSync();
-    final certString = base64Encode(bytes);
-    return certString;
+  Result<String, PixException> certFileToBase64String(
+      {required File pkcs12CertificateFile}) {
+    try {
+      final bytes = pkcs12CertificateFile.readAsBytesSync();
+      final certString = base64Encode(bytes);
+      return Success(certString);
+    } on PathNotFoundException catch (e) {
+      final error = SicoobCertificateException.pathNotFoundException(e);
+      return Failure(error);
+    } catch (e) {
+      final unknownError = SicoobUnknownException.unknownException(e);
+      return Failure(unknownError);
+    }
   }
 
-  /// Converts a base64-encoded certificate string to a byte array, for use in secure communication.
+  /// Returns a [Result] object containing a byte array on success or a [PixException]
+  /// on failure. The input is a base64-encoded string that is first decoded into a
+  /// byte array.
   ///
-  /// Returns a [Result] object that contains the byte array, or a [PixError] if the string is invalid.
-  ///
-  /// Example usage:
-  ///
-  /// ```dart
-  /// final security = SicoobSecurity();
-  /// final certString = 'base64EncodedString';
-  /// final bytesResult = security.certificateStringToBytes(certString);
-  /// final bytes = bytesResult.getOrThrow();
-  ///   // Use the byte array for secure communication
-  /// ```
+  /// Throws a [SicoobCertificateException] if there is an issue with the certificate,
+  /// or a [SicoobUnknownException] for any other type of error
   @override
-  Result<Uint8List, PixError> certificateStringToBytes(
+  Result<Uint8List, PixException> certificateStringToBytes(
       String certificateBase64String) {
     try {
       var certificateBytes = base64.decode(certificateBase64String);
       return Success(certificateBytes);
     } on FormatException catch (e) {
-      if (e.message.contains('Invalid length, must be multiple of four')) {
-        return Failure(PixError('invalid_certificate_base64String'));
-      } else {
-        return Failure(PixError(e.toString()));
-      }
+      final error = SicoobCertificateException.formatException(e);
+      return Failure(error);
+    } catch (e) {
+      final unknownError = SicoobUnknownException.unknownException(e);
+      return Failure(unknownError);
     }
   }
 }