+++ test/src/features/token/repository/token_repository_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:pix_sicoob/src/error/pix_error.dart';
+import 'package:pix_sicoob/src/errors/sicoob_api_exception.dart';
 import 'package:pix_sicoob/src/features/token/model/token.dart';
 import 'package:pix_sicoob/src/features/token/repository/token_repository.dart';
 import 'package:pix_sicoob/src/services/client_service.dart';
@@ -12,7 +12,6 @@ void main() {
   late ClientService client;
   late TokenRepository tokenRepository;
   late Uri uri;
-  late dynamic expectedResponse;
   const clientID = 'client-id';
   setUp(() {
     client = MockClientService();
@@ -21,58 +20,54 @@ void main() {
     uri = Uri.parse('https://teste.example.com/auth');
     registerFallbackValue(uri);
   });
-  test('A chamada da API deve retornar um Token', () async {
-    // Configura as expectativas do teste
-    expectedResponse = {
-      'access_token': '234k-vkhk-l1h2-3l3l',
-      'expires_in': 20,
-      'refresh_expires_in': 300,
-      'token_type': 'Bearer',
-      'not-before-policy': 30,
-      'scope': 'pix.read',
-    };
 
-    // Configura o mock da chamada http
-    when(() => client.post(
-          any(),
-          body: any(named: 'body'),
-          headers: any(named: 'headers'),
-        )).thenAnswer((_) async => Success(expectedResponse));
+  group('Token: ', () {
+    test('A chamada da API deve retornar um Token', () async {
+      final expectedResponse = {
+        'access_token': '234k-vkhk-l1h2-3l3l',
+        'expires_in': 20,
+        'refresh_expires_in': 300,
+        'token_type': 'Bearer',
+        'not-before-policy': 30,
+        'scope': 'pix.read',
+      };
 
-    // Executa o método que está sendo testado
-    final response = await tokenRepository.getToken(
-      uri: uri,
-      clientID: clientID,
-    );
-    var result = response.fold((success) => success, (failure) => null);
+      when(() => client.post(any(),
+              body: any(named: 'body'), headers: any(named: 'headers')))
+          .thenAnswer((_) async => Success(expectedResponse));
 
-    // Verifica se o resultado foi o esperado
-    expect(result, isNotNull);
-    expect(result, isA<Token>());
-    expect(result?.accessToken, equals(expectedResponse['access_token']));
-  });
+      final response = await tokenRepository.getToken(
+        uri: uri,
+        clientID: clientID,
+      );
 
-  test('Em caso de erro a chamada da API deve retornar um PixError', () async {
-    // Configura as expectativas do teste
-    expectedResponse = "error";
+      var result = response.getOrThrow();
 
-    // Configura o mock da chamada http
-    when(() => client.post(
-          any(),
-          body: any(named: 'body'),
-          headers: any(named: 'headers'),
-        )).thenAnswer((_) async => Failure(PixError(expectedResponse)));
+      expect(result, isNotNull);
+      expect(result, isA<Token>());
+      expect(result.accessToken, equals(expectedResponse['access_token']));
+    });
 
-    //Executa o metodo que está sendo testado
-    final response = await tokenRepository.getToken(
-      uri: uri,
-      clientID: clientID,
-    );
-    var result = response.fold((success) => null, (failure) => failure)!;
+    test(r'''Map da response não contem uma key "access_token":
+              O Metodo deverá retornar uma failure de SicoobApiException.
+          ''', () async {
+      final expectedResponse = {'error': 'type'};
 
-    // Verifica se o resultado foi o esperado
-    expect(result, isA<Exception>());
-    expect(result.message, isNotNull);
-    expect(result.message, contains('error'));
+      when(() => client.post(
+            any(),
+            body: any(named: 'body'),
+            headers: any(named: 'headers'),
+          )).thenAnswer((_) async => Success(expectedResponse));
+
+      final response = await tokenRepository.getToken(
+        uri: uri,
+        clientID: clientID,
+      );
+      var result = response.exceptionOrNull();
+
+      expect(result, isA<SicoobApiException>());
+      expect(result!.message, isNotNull);
+      expect(result.message, contains('error'));
+    });
   });
 }