+++ test/src/shared/services/sicoob_security_test.dart
@@ -1,29 +1,30 @@
 import 'dart:io';
+import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:pix_sicoob/src/error/pix_error.dart';
+import 'package:pix_sicoob/src/errors/sicoob_certificate_exception.dart';
 import 'package:pix_sicoob/src/services/client_security.dart';
 import 'package:pix_sicoob/src/services/sicoob_security.dart';
 
 void main() {
-  late ClientSecurity sicoobSecurity;
+  late ClientSecurity clientSecurity;
   late String certificateBase64String;
   const certificatePassword = '1234';
 
   setUp(() {
-    sicoobSecurity = SicoobSecurity();
-    certificateBase64String = sicoobSecurity.certFileToBase64String(
-      pkcs12CertificateFile: File('test/cert/cert_for_test.pfx'),
-    );
+    clientSecurity = SicoobSecurity();
+    certificateBase64String = clientSecurity
+        .certFileToBase64String(
+            pkcs12CertificateFile: File('test/cert/cert_for_test.pfx'))
+        .getOrThrow();
   });
 
-  test('A Conversao do certificado deve retornar uma String', () {
-    // Verifica se o resultado foi o esperado
-    expect(certificateBase64String, isA<String>());
-  });
+  group('Security Context: ', () {
+    test('A Conversao do certificado deve retornar uma String', () {
+      expect(certificateBase64String, isA<String>());
+    });
 
-  group('SecurityContext', () {
     test('O Metodo deve retornar um Security Context', () {
-      final result = sicoobSecurity.getContext(
+      final result = clientSecurity.getContext(
         certificateBase64String: certificateBase64String,
         certificatePassword: certificatePassword,
       );
@@ -31,32 +32,72 @@ void main() {
       expect(securityContext, isA<SecurityContext>());
     });
 
-    test('Deve retornar um PixError se a senha for inválida', () {
-      // Executa o método que está sendo testado
-      final response = sicoobSecurity.getContext(
+    test(r'''Caso a senha do certificado esteja incorreta:
+    Deverá retornar um SicoobCertificateException do tipo: incorrectCertificatePassword''',
+        () {
+      final response = clientSecurity.getContext(
         certificateBase64String: certificateBase64String,
         certificatePassword: '123',
       );
 
-      // Verifica se o resultado foi o esperado
       final result = response.exceptionOrNull();
-      expect(result, isA<PixError>());
-      expect(result!.message, isNotNull);
-      expect(result.message.values, contains('incorrect_certificate_password'));
+      expect(result, isA<SicoobCertificateException>());
+      expect(result!.exceptionType,
+          equals(CertificateExceptionType.incorrectCertificatePassword));
     });
 
-    test('Deve retornar um PixError se o certificado for inválido', () {
-      // Executa o método que está sendo testado
-      final response = sicoobSecurity.getContext(
+    test(r'''Caso o certificado seja invalido:
+    Deverá retornar um SicoobCertificateException do tipo: invalidPkcs12Certificate''',
+        () {
+      final response = clientSecurity.getContext(
         certificateBase64String: '${certificateBase64String}1123',
         certificatePassword: '1234',
       );
 
-      // Verifica se o resultado foi o esperado
       final result = response.exceptionOrNull();
-      expect(result, isA<PixError>());
-      expect(result!.message, isNotNull);
-      expect(result.message.values, contains('invalid_pkcs12_certificate'));
+      expect(result, isA<SicoobCertificateException>());
+      expect(
+        result!.exceptionType,
+        equals(CertificateExceptionType.invalidPkcs12Certificate),
+      );
+    });
+
+    test(r'O metodo deve retornar uma uma lista de Bytes Uint8List', () {
+      final response = clientSecurity.certificateStringToBytes(
+        certificateBase64String,
+      );
+
+      final result = response.getOrThrow();
+      expect(result, isA<Uint8List>());
+      expect(result, isNotEmpty);
+    });
+
+    test(r'''Caso o certificado em base64String seja invalido:
+    Deverá retornar um SicoobCertificateException do tipo: invalidCertificateBase64String''',
+        () {
+      final response = clientSecurity.certificateStringToBytes(
+        '$certificateBase64String unk',
+      );
+
+      final result = response.exceptionOrNull();
+      expect(result, isA<SicoobCertificateException>());
+      expect(
+        result!.exceptionType,
+        equals(CertificateExceptionType.invalidCertificateBase64String),
+      );
+    });
+
+    test(r'''Caso certificado não for encontrado:
+    Deverá retornar um SicoobCertificateException do tipo: certificateFilePathNotFound''',
+        () {
+      final response = clientSecurity.certFileToBase64String(
+          pkcs12CertificateFile: File('invalidPath'));
+      final result = response.exceptionOrNull();
+      expect(result, isA<SicoobCertificateException>());
+      expect(
+        result!.exceptionType,
+        equals(CertificateExceptionType.certificateFilePathNotFound),
+      );
     });
   });
 }