+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.1.0] - 24-03-2023
+
+- Improved unit tests
+- Created classes for specific error handling
+  1. Base class for all exceptions related to the Pix feature.
+  2. An exception representing an error response from the Sicoob API.
+  3. Exception thrown when an error occurs with the certificate used for Pix
+  4. Exception thrown when an HTTP request fails.
+  5. An exception that represents an unknown error in the Pix Sicoob API.
+
 ## [1.0.2] - 22-03-2023
 
 - Updated readme "How To Use"